+++ omechu-app/src/app/(auth)/sign-in/layout.tsx
@@ -3,7 +3,7 @@
 import Image from "next/image";
 import { useRouter } from "next/navigation";
 
-import { Header } from "@/shared/ui/Header";
+import { Header } from "@/shared/ui/header/Header";
 
 /**
  * 로그인 관련 페이지 레이아웃
