--- omechu-app/src/shared/ui/auth/index.ts
@@ -1,1 +1,0 @@
-export { AuthButton } from "./AuthButton";
