+++ omechu-app/src/shared/ui/progress/ProgressBar.tsx
