--- omechu-app/src/shared/ui/terms/index.ts
@@ -1,1 +1,0 @@
-export { default as TermsContent } from "./TermsContent";
